perf: remove changeMenu, perf improe toast, a11y
