+++ components/cookies/index.tsx
@@ -1,20 +1,21 @@
+import CookieContextProvider from "@/context/cookie";
 import CookieFormContextProvider from "@/context/cookie/form";
 import { cn } from "@/lib/utils";
 
-import ChangeMenu from "./changeMenu";
 import ManageConsent from "./manage";
 import CookieToast from "./toast";
 
 export default function CookieTray() {
   return (
     <>
-      <div className={cn("transition-all")}>
-        <CookieFormContextProvider>
-          <ManageConsent />
-        </CookieFormContextProvider>
-        <ChangeMenu />
-        <CookieToast />
-      </div>
+      <CookieContextProvider>
+        <div className={cn("transition-all")}>
+          <CookieFormContextProvider>
+            <ManageConsent />
+          </CookieFormContextProvider>
+          <CookieToast />
+        </div>
+      </CookieContextProvider>
     </>
   );
 }
